[Function] Define new Function Value.
This function value will refer to the function object in the stack, in
practice represents a compiled function that stands as a stack frame
also in the vm execution stack.
A function contains a name and a chunk of code and, possibly, other
child/sub functions inside it.
When we compile the main AST Program object, we emit a Function with
name "<main>" that will represent the top-level function for execution,
this function will contains other functions as well.
For the moment there is function compilation without function invocation
and locals. The locals will be reserved on the main vm stack at
invokation time, in that moment I'll add the argument binding before
switching the current chunk for the function chunk and pushing the vm ip
for continuation after function return, typical stuff.