[VM][Function] Implement simple function call.
Function calls without arguments and return values. The mechanism is
that we have a call stack with frames and we push a new frame with a
function and an IP pointer. The interpreter loop executes the opcodes of
the current frame that is the frame at the top of the stack.
Need to implement support to pop a frame and resume execution on the
previous frame.