[Locals] Implementation of local variables.
Implemented local variables using the stack, the locals are not in a
symbol table, but instead are stored in the stack following the function
frame. When a function returns the locals will be removed by shrinking
the stack to the base ptr of the frame.
For lexical scoping the locals are just indices into the stack, also
when we are in the global lexical scope or we are in the first frame
function, the function with the name "<main>" we check if the variable
is refering to a global variable and we emit GetGlobal Opcode instead.
Global variables are stored in the symbol table in the heap.