Initial Impliementation of Native Functions.
Initial implementation of native functions calls. These "native"
functions are implemented directly on the interpreter in Rust, thats why
we call it "native". All natives functions have the same signature, the
arguments is a vector of Object and the result is one Object. If in the
future we want to support returning multiple values, we just return a
Tuple Object. Implemented two functions, abs and print. We only support
abs in the VM logic, in next commit, I'll remove the print statement as
OpCode and implement it as a function instead.