Removed Print Statement and OpCode.
I'm converting the print statement into a native function call, so first
step is to remove the statement from the lexer, parser and bytecode
spec. Then I'll add support to call the print function, for the moment
the print function will only support 1 argument because we don't have
variadic function support in this interpreter.