[Compiler][VM][Function] Support variadic functions.
When a function was a arity (arg count) equal to usize::MAX, we consider
the function to receive a variadic number of arguments, so we don't do
any argument count check and we just forward the arguments specified
during the call expression to the function. There is no check for
minimum number of variadic function arguments.